Fix 26.1 compile errors and implement the ClickGUI, HUD, and Entity ESP
Build fixes (now compiles clean against MC 26.1.2 / Java 25):
- AutoPearl: ThrownEnderpearl moved to the projectile.throwableitemprojectile package
- MaceCombat: Entity.fallDistance is now double — cast to float

New UI, built on verified 26.1 render APIs (GuiGraphicsExtractor /
HudElementRegistry / LevelRenderEvents — all stable Fabric surfaces, no mixins):
- BotScreen: Right-Shift ClickGUI grouping modules by category; click a row to
  toggle (green = on), hover tooltips, auto-column wrapping
- HudOverlay: top-right arraylist of enabled modules + watermark, hidden while a
  screen or the F3 debug overlay is open
- EntityESP: per-frame colored outline boxes (foe=red, friend=green, neutral
  player=white, hostile=orange, other=yellow) via LevelRenderEvents
- Wired HUD + ESP registration into MCBotClient

Also: generated the missing mod icon (assets/mcbot/icon.png) and brought the
README up to date (#bot prefix, Java 25, Meteor Baritone, new modules/keybinds).

+++ src/main/java/com/mcbot/module/combat/MaceCombatModule.java
@@ -158,7 +158,7 @@ public class MaceCombatModule extends Module {
         EntityUtil.lookAt(client, target);
 
         if (CombatUtil.isCooldownReady(client, 0.85f)) {
-            float fall = client.player.fallDistance;
+            float fall = (float) client.player.fallDistance;
             CombatUtil.attack(client, target);
             float dmg = CombatUtil.estimateMaceDamage(fall);
             announce(client, String.format("§a[MC BOT] Mace SMASH! ~%.0f dmg from %.1f blocks", dmg, fall));
